modify: #38 전체 카테고리 리스트 조회 API 수정

+++ src/main/java/com/waguwagu/weat/domain/category/model/entity/Category.java
@@ -26,5 +26,5 @@ public class Category {
 
     @Column(name = "category_order")
     @Comment("카테고리순서")
-    private Integer categoryOrder;
+    private Long categoryOrder;
 }
